requirements.txt
streamlit==1.50.0
pandas==2.2.3
numpy==2.1.3
pyarrow==17.0.0
scikit-learn==1.7.2
xgboost==3.0.3
lightgbm==4.6.0
catboost==1.2.5
optuna==3.6.1
imbalanced-learn==0.12.4
plotly==5.24.1
matplotlib==3.9.2
loguru==0.7.2
python-dotenv==1.0.1
openai==1.47.0
tiktoken==0.7.0
mlflow==2.16.2
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.35.21
s3fs==2024.6.0
psycopg2-binary==2.9.9
redis==5.0.8
rq==1.16.2
weasyprint==65.0
kaleido==0.2.1
requests==2.32.3
